core: Move copy_on_cpu() to a separate file
operations.rs is pretty large and optimizing and fixing copy_on_cpu()
will produce more code and functions, so move it to a separate file for
better readability and encapsulation.
